// src/app/(system)/(service)/adminUser/types.ts
export type GetPermissionTreeResponse = {
  id: string;
  level: string;
  type: string;
  code: string;
  parentCode: string;
  displayName: string;
  children: GetPermissionTreeResponse;
}[];